Create 2lab.cpp
N people numbered from 0 to N - 1 participate in the tournament. Each player is represented by a node in the graph.

You are given an integer N and a two-dimensional integer array of edges of length M representing the graph, where edge [i] = [u, v] indicates that there is a directed edge from player u to player v in the graph.

A directed edge from u to v in the graph means that player u is stronger than player v and player v is weaker than player u.

Player u will become the champion of the tournament if there is no player v who is stronger than player u. At the same time, player u cannot become the champion of the tournament if there is player k for whom there are no stronger players and it is impossible to determine which of players u and k is the stronger player.

Return the number of the player who will become the tournament champion if such a player exists, otherwise return -1.